+++ libvtkrs/src/vtk_object.cpp
@@ -1,6 +1,7 @@
 #include "vtk_object.h"
 #include "vtk_object.rs.h"
 
+#include <cstdint>
 #include <vtkObject.h>
 
 using vtkObject = vtkObject;
@@ -27,11 +28,11 @@ void modified(const vtkObject &obj) {
     obj2.Modified();
 }
 
-void remove_observer(vtkObject &obj, unsigned long tag) {
+void remove_observer(vtkObject &obj, uint64_t tag) {
     obj.RemoveObserver(tag);
 }
 
-void remove_observers(vtkObject &obj, unsigned long tag) {
+void remove_observers(vtkObject &obj, uint64_t tag) {
     obj.RemoveObservers(tag);
 }
 
@@ -39,7 +40,7 @@ void remove_all_observer(vtkObject &obj) {
     obj.RemoveAllObservers();
 }
 
-int has_observer(const vtkObject &obj, unsigned long event) {
-    vtkObject &obj2 = const_cast<vtkObject&>(obj);
+uint64_t has_observer(const vtkObject &obj, uint64_t event) {
+    vtkObject &obj2 = const_cast<vtkObject &>(obj);
     return obj2.HasObserver(event);
 }
